set up APIRouter
