+++ .vscode/launch.json
@@ -1,7 +1,4 @@
 {
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
         {
@@ -17,6 +14,8 @@
             ],
             "jinja": true,
             "justMyCode": true,
+            "console": "integratedTerminal",
+            "subProcess": true,
             "preLaunchTask": "Open in browser"
         }
     ]
